requirements.txt
playwright>=1.40
python-dotenv>=1.0
pyyaml>=6.0
flask>=3.0
